lua folder structure, custom headers, OpenId Connect access tokens

+++ t/run-flow.sh
@@ -15,10 +15,24 @@ fi
 
 echo "Building test runner Docker image (please be patient first time)..."
 TEST_RUNNER_IMAGE_ID="$(docker build -q $DIR)"
+if [ -z "$TEST_RUNNER_IMAGE_ID" ]
+then
+    echo "test runner image build error"
+    exit 1
+fi
+echo "Done"
+
+echo "Build gluu-gateway docker image"
+export GG_IMAGE_ID="$(docker build -q -f $HOST_GIT_ROOT/Dockerfile $HOST_GIT_ROOT)"
+if [ -z "$GG_IMAGE_ID" ]
+then
+    echo "gluu-gateway image build error"
+    exit 1
+fi
 echo "Done"
 
 docker run --net host --rm -v /var/run/docker.sock:/var/run/docker.sock \
     -v $HOST_GIT_ROOT:$GIT_ROOT \
     -v /tmp:/tmp \
-    --env HOST_GIT_ROOT --env GIT_ROOT \
-    $TEST_RUNNER_IMAGE_ID busted -m=$GIT_ROOT/t/lib/?.lua $GIT_ROOT/t/flow$TEST
+    --env HOST_GIT_ROOT --env GIT_ROOT --env GG_IMAGE_ID \
+    $TEST_RUNNER_IMAGE_ID busted -m=$GIT_ROOT/t/lib/?.lua $GIT_ROOT/t/specs$TEST
